Update tmr_demo example to autoconvert graphics
This one was tricky.

The files with the graphics data have been edited by hand. A lot. Even
adding data structures and typedefs there (which were copied from
libtonc anyway).

Also, there doesn't seem to be any option for grit to not export the
first tile as an empty tile, so it was needed to update the code.

Finally, the background color of the original berk.bmp file isn't the
same as the one in the converted graphics. I've created a new version of
the file with the actual background color used in the demo.

The old berk.bmp file has been preserved for reference.