Convert graphics of txt_se1 example at build time
This has been a bit annoying. The example assumes that the order of the
palette of the border is a specific one, which shouldn't be assumed.

I've decided to fix that by not printing text with the palette of the
border.

Also, palette 14 was generated using the border palette as base, which
isn't reliable either, so now it's generated by hand.